fix(http-display types): Move http inputs type to correct file
