+++ src/pages/main-page/components/http-display/components/http-request-form/types.ts
@@ -1,23 +1,4 @@
-import { InterfaceType } from '../../../../../../common/types/api-interface-types';
-
-export enum HTTPRequestMethod {
-  GET = 'GET',
-  POST = 'POST',
-  PUT = 'PUT',
-  PATCH = 'PATCH',
-  DELETE = 'DELETE',
-  OPTIONS = 'OPTIONS',
-}
-
 export enum HttpRequestFormTab {
   HEADERS = 'headers',
   BODY = 'body',
-}
-
-export interface HTTPInputs {
-  interfaceType: InterfaceType.HTTP;
-  url: string;
-  method: HTTPRequestMethod;
-  headers: { name: string; value: string }[];
-  body: string;
 }
