+++ src/pages/main-page/components/http-display/http-display.tsx
@@ -1,5 +1,4 @@
 import { SubmitHandler, UseFormReturn } from 'react-hook-form';
-import { HTTPInputs } from './components/http-request-form/types';
 import { Box } from '@mui/material';
 import { GoGrabber } from 'react-icons/go';
 import { PanelGroup, Panel, PanelResizeHandle } from 'react-resizable-panels';
@@ -8,9 +7,10 @@ import HTTPResponse from './components/http-response/http-response';
 import { useState } from 'react';
 import { InterfaceInputs } from '../../types';
 import { InterfaceType } from '../../../../common/types/api-interface-types';
+import { HTTPInputs } from './types';
 
 type HTTPDisplayProps = {
-  interfaceForm: UseFormReturn<HTTPInputs, any, HTTPInputs>;
+  interfaceForm: UseFormReturn<HTTPInputs>;
   onRecord: (record: InterfaceInputs) => void;
   onSaveRecord: (record: InterfaceInputs) => void;
 };
