+++ src/pages/main-page/types.ts
@@ -1,4 +1,4 @@
-import { HTTPInputs } from './components/http-display/components/http-request-form/types';
+import { HTTPInputs } from './components/http-display/types';
 import { WSInputs } from './components/ws-display/types';
 
 export type InterfaceInputs = HTTPInputs | WSInputs;
