+++ src/pages/main-page/utils.ts
@@ -1,5 +1,5 @@
 import { InterfaceType } from '../../common/types/api-interface-types';
-import { HTTPInputs } from './components/http-display/components/http-request-form/types';
+import { HTTPInputs } from './components/http-display/types';
 import { WSInputs } from './components/ws-display/types';
 import { InterfaceInputs } from './types';
